tests/wpt/web-platform-tests/webmessaging/with-ports/023.html

<!doctype html>
<title>null ports</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
async_test(function(t) {
  postMessage('', '*', null);
  onmessage = t.step_func(function(e) {
    assert_array_equals(e.ports, []);
    this.done();
  });
});
</script>
